Refactor: Extract Script component from App to support multiple scripts
